+++ plugins/sidepad/hooks/pointer/index.ts
@@ -4,6 +4,7 @@ export * from './code-pointer-step';
 export * from './code-window';
 export * from './dragged-range-of';
 export * from './edge-tick';
+export * from './listen-for-drag';
 export * from './page-edge-tick';
 export * from './page-pointer-step';
 export * from './page-window';